git-db print suffix and database
This was changed to support pbak (backup postgres database)

It has been hardcoded to work with vmdb database. Not sure the best way to derive the base name
besides introducing yet another git configuration variable.
It was picking up any database with the name development in it - which is basically any rails
Since I only care about vmdb right now, it works for me

tried to make the shape of the code more consistent. Now run handles params and prints.